requirements.txt
PyGithub>=2.1.1
python-dotenv>=1.0.0
tqdm>=4.66.1
pytz>=2023.3
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.90.0
langchain>=0.2.0
langchain-openai>=0.1.0
browser-use>=0.1.0
asyncio

# Development dependencies
flake8>=7.0.0
mypy>=1.9.0
black>=24.3.0
isort>=5.13.0
